Normalize all files in a directory simultaneously
Instead of processing one file at a time, process all files in a
directory at the same time. This may use all available CPU if a
directory has a lot of files in it. A possible fix for this would be to
launch each normalize process with a niceness to yield to more important
tasks...