Remove any ID3 tags
If ID3 tags are present, flac reencoding will fail with an error message
similar to this,

ERROR: input file <file> has an ID3v2 tag

To get rid of this stale metadata, the flac file will need to be fully
decoded to a wave file and reencoded back to flac from that. This step
will not preserve metadata, so the original metadata has to be exported
and imported back. This is possible only for the text metadata. So, the
album art is re-embedded. The original file may not have had any replay
gain data. So, to be safe, any replay gain data is dropped and the
replay gain is freshly calculated.