Fix issue with spaces in a file name
When a single FLAC file is given as the argument, spaces in the file
name caused the script to go crazy. This is a bash/getopts limitation
and other than writing my own argument parser, there's no good fix.

This effectively means that the script can take exactly one argument,
which is either a single FLAC file or a directory. No argument is
interpreted as the current directory. Any more than one argument will
break things.

This makes this script a prime candidate to move to ruby or python.