Rewriting script using ruby
Since the bash script has issues handling multiple inputs with spaces in
the file name, it's being rewritten in ruby.

This first commit records the completion of the basic FlacFile class
which deals with normalizing and/or re-encoding a single FLAC file. If
the file is not a valid FLAC file, it is left alone.