Multi-threaded directory recurser function done
This function will enter a directory and fork off processes that will
simultaneously process all content in that directory.

Now, this only really expects to find all FLAC files at the second
directory level, so this may cause issues if you have FLAC files at
different levels. In that case, may be forking may not be a great
idea...